app/requirements.txt
discord.py==2.3.2
aiohttp==3.13.2
python-dotenv==1.2.1
openai==1.68.1
dnspython==2.8.0
pymongo==4.15.5
fastapi
uvicorn
